ia_reportes/requirements.txt
openai==1.3.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
pandas==2.1.0
openpyxl==3.1.0
reportlab==4.0.7
requests==2.31.0
matplotlib==3.8.0
seaborn==0.13.0
